backend/requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
flask-migrate
psycopg2-binary
flask-cors